@@ -1,28 +1,28 @@
 {
-  "_from": "mysql",
-  "_id": "mysql@2.15.0",
+  "_from": "mysql@^2.15.0",
+  "_id": "mysql@2.17.1",
   "_inBundle": false,
-  "_integrity": "sha512-C7tjzWtbN5nzkLIV+E8Crnl9bFyc7d3XJcBAvHKEVkjrYjogz3llo22q6s/hw+UcsE4/844pDob9ac+3dVjQSA==",
+  "_integrity": "sha512-7vMqHQ673SAk5C8fOzTG2LpPcf3bNt0oL3sFpxPEEFp1mdlDcrLK0On7z8ZYKaaHrHwNcQ/MTUz7/oobZ2OyyA==",
   "_location": "/mysql",
   "_phantomChildren": {},
   "_requested": {
-    "type": "tag",
+    "type": "range",
     "registry": true,
-    "raw": "mysql",
+    "raw": "mysql@^2.15.0",
     "name": "mysql",
     "escapedName": "mysql",
-    "rawSpec": "",
+    "rawSpec": "^2.15.0",
     "saveSpec": null,
-    "fetchSpec": "latest"
+    "fetchSpec": "^2.15.0"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/mysql/-/mysql-2.15.0.tgz",
-  "_shasum": "ea16841156343e8f2e47fc8985ec41cdd9573b5c",
-  "_spec": "mysql",
-  "_where": "/Users/jmoussa/server",
+  "_resolved": "https://registry.npmjs.org/mysql/-/mysql-2.17.1.tgz",
+  "_shasum": "62bba4a039a9b2f73638cd1652ce50fc6f682899",
+  "_spec": "mysql@^2.15.0",
+  "_where": "/Users/jmoussa/code_react/money_manager/server",
   "author": {
     "name": "Felix Geisendörfer",
     "email": "felix@debuggable.com",
@@ -55,19 +55,18 @@
     }
   ],
   "dependencies": {
-    "bignumber.js": "4.0.4",
-    "readable-stream": "2.3.3",
-    "safe-buffer": "5.1.1",
-    "sqlstring": "2.3.0"
+    "bignumber.js": "7.2.1",
+    "readable-stream": "2.3.6",
+    "safe-buffer": "5.1.2",
+    "sqlstring": "2.3.1"
   },
   "deprecated": false,
   "description": "A node.js driver for mysql. It is written in JavaScript, does not require compiling, and is 100% MIT licensed.",
   "devDependencies": {
     "after": "0.8.2",
-    "eslint": "4.8.0",
-    "nyc": "10.3.2",
-    "seedrandom": "2.4.3",
-    "timezone-mock": "0.0.5",
+    "eslint": "5.15.1",
+    "seedrandom": "3.0.1",
+    "timezone-mock": "0.0.7",
     "urun": "0.0.8",
     "utest": "0.0.8"
   },
@@ -91,9 +90,9 @@
   "scripts": {
     "lint": "eslint .",
     "test": "node test/run.js",
-    "test-ci": "nyc --reporter=text npm test",
-    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "test-ci": "node tool/install-nyc.js --nyc-optional --reporter=text -- npm test",
+    "test-cov": "node tool/install-nyc.js --reporter=html --reporter=text -- npm test",
     "version": "node tool/version-changes.js && git add Changes.md"
   },
-  "version": "2.15.0"
+  "version": "2.17.1"
 }
