@@ -1,27 +1,27 @@
 {
-  "_from": "sqlstring@2.3.0",
-  "_id": "sqlstring@2.3.0",
+  "_from": "sqlstring@2.3.1",
+  "_id": "sqlstring@2.3.1",
   "_inBundle": false,
-  "_integrity": "sha1-UluKT9Jtb3GqYegipsr5dtMa0qg=",
+  "_integrity": "sha1-R1OT/56RR5rqYtyvDKPRSYOn+0A=",
   "_location": "/sqlstring",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "sqlstring@2.3.0",
+    "raw": "sqlstring@2.3.1",
     "name": "sqlstring",
     "escapedName": "sqlstring",
-    "rawSpec": "2.3.0",
+    "rawSpec": "2.3.1",
     "saveSpec": null,
-    "fetchSpec": "2.3.0"
+    "fetchSpec": "2.3.1"
   },
   "_requiredBy": [
     "/mysql"
   ],
-  "_resolved": "https://registry.npmjs.org/sqlstring/-/sqlstring-2.3.0.tgz",
-  "_shasum": "525b8a4fd26d6f71aa61e822a6caf976d31ad2a8",
-  "_spec": "sqlstring@2.3.0",
-  "_where": "/Users/jmoussa/server/node_modules/mysql",
+  "_resolved": "https://registry.npmjs.org/sqlstring/-/sqlstring-2.3.1.tgz",
+  "_shasum": "475393ff9e91479aea62dcaf0ca3d14983a7fb40",
+  "_spec": "sqlstring@2.3.1",
+  "_where": "/Users/jmoussa/code_react/money_manager/server/node_modules/mysql",
   "bugs": {
     "url": "https://github.com/mysqljs/sqlstring/issues"
   },
@@ -58,7 +58,7 @@
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "4.8.0",
+    "eslint": "4.18.1",
     "eslint-plugin-markdown": "1.0.0-beta.6",
     "nyc": "10.3.2",
     "urun": "0.0.8",
@@ -94,5 +94,5 @@
     "test-ci": "nyc --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.3.0"
+  "version": "2.3.1"
 }
